fix(recipe): deleted the image when deleting the recipe

+++ src/controllers/recipe.controller.ts
@@ -1,43 +1,43 @@
-import { Recipe, User }                 from "@prisma/client";
-import { type Request, type Response  } from "express";
-import { recipeService }                from "../services/recipe.service";
-import { 
-  ICreateRecipeSchema, 
-  ICreateStepSchema, 
-  IUpdateRecipeSchema, 
+import { Recipe, User } from "@prisma/client";
+import { type Request, type Response } from "express";
+import { recipeService } from "../services/recipe.service";
+import {
+  ICreateRecipeSchema,
+  ICreateStepSchema,
+  IUpdateRecipeSchema,
   IUpdateStepSchema
-}                                       from "../schemas/recipe.schema";
-import { RecipePreviewDTO }             from "../dtos/recipe.dto";
-import { verifyToken }                  from "../utils/token";
-import { stepService }                  from "../services/step.service";
-import { likeService }                  from "../services/like.service";
-import { commentService }               from "../services/comment.service";
+} from "../schemas/recipe.schema";
+import { RecipePreviewDTO } from "../dtos/recipe.dto";
+import { verifyToken } from "../utils/token";
+import { stepService } from "../services/step.service";
+import { likeService } from "../services/like.service";
+import { commentService } from "../services/comment.service";
 import { DeleteObjectsCommand } from "@aws-sdk/client-s3";
 import { s3 } from "../configs/aws.config";
 
 class RecipeController {
   public async create(request: Request, response: Response) {
-    const user   = request.user as User;
-    const data   = request.body as ICreateRecipeSchema;
+    const user = request.user as User;
+    const data = request.body as ICreateRecipeSchema;
     const recipe = await recipeService.createRecipe({ ...data, ownerId: user.id });
 
     response.send(recipe);
   };
 
   public async getRecipe(request: Request, response: Response) {
-    const recipeId    = request.params.recipeId as string;
-    const recipe      = await recipeService.getRecipeById(recipeId);
+    const recipeId = request.params.recipeId as string;
+    const recipe = await recipeService.getRecipeById(recipeId);
     const accessToken = request.headers.authorization;
 
     if (recipe === null) {
       return response.status(404).send({
-        code   : "recipe-not-found",
+        code: "recipe-not-found",
         message: "Recipe not found!"
       });
     }
 
     if (accessToken) {
-      const userId           = verifyToken(accessToken);
+      const userId = verifyToken(accessToken);
       const isAlreadyVisited = await recipeService.getVisitedRecipeByIds(recipe.id, userId);
 
       if (isAlreadyVisited === null) {
@@ -50,7 +50,7 @@ class RecipeController {
 
   public async update(request: Request, response: Response) {
     const recipeId = request.params.recipeId as string;
-    const data     = request.body as IUpdateRecipeSchema;
+    const data = request.body as IUpdateRecipeSchema;
 
     const recipe = await recipeService.getRecipeById(recipeId);
 
@@ -93,21 +93,39 @@ class RecipeController {
 
     if (recipe === null) {
       return response.status(404).send({
-        code   : "recipe-not-found",
+        code: "recipe-not-found",
         message: "Recipe not found!",
       });
     }
-    
-    await recipeService.deleteRecipe(recipeId);
 
-    response.send({ message: "Recipe deleted!" });
-  };
+    if (recipe.image !== '') {
+      const objectsToDelete = [];
+
+      objectsToDelete.push({ Key: (recipe.image).split("/").pop() });
+
+      if (objectsToDelete.length > 0) {
+        const params = {
+          Bucket: 'culinarybook-images',
+          Delete: {
+            Objects: objectsToDelete
+          }
+        };
+
+        const command = new DeleteObjectsCommand(params);
+        await s3.send(command);
+      }
+
+      await recipeService.deleteRecipe(recipeId);
+
+      response.send({ message: "Recipe deleted!" });
+    };
+  }
 
   public async getMyRecipes(request: Request, response: Response) {
-    const user    = request.user as User;
-    const sortBy  = request.query.sortBy as 'asc' | 'desc';
-    const page    = parseInt(request.query.page as string) || 1;
-    const limit   = parseInt(request.query.limit as string ) || 10;
+    const user = request.user as User;
+    const sortBy = request.query.sortBy as 'asc' | 'desc';
+    const page = parseInt(request.query.page as string) || 1;
+    const limit = parseInt(request.query.limit as string) || 10;
 
     const recipes = await recipeService.getMyRecipesByUserId(user.id, sortBy, page, limit);
 
@@ -116,9 +134,9 @@ class RecipeController {
   };
 
   public async getLikedRecipes(request: Request, response: Response) {
-    const user  = request.user as User;
-    const page  = parseInt(request.query.page as string) || 1;
-    const limit = parseInt(request.query.limit as string ) || 10;
+    const user = request.user as User;
+    const page = parseInt(request.query.page as string) || 1;
+    const limit = parseInt(request.query.limit as string) || 10;
 
     const recipes = await recipeService.getLikedRecipesByUserId(user.id, page, limit);
 
@@ -126,31 +144,31 @@ class RecipeController {
   };
 
   public async getRecommendedRecipes(request: Request, response: Response): Promise<void> {
-    const user       = request.user as User;
-    const page       = parseInt(request.query.page as string) || 1;
-    const limit      = parseInt(request.query.limit as string ) || 10;
+    const user = request.user as User;
+    const page = parseInt(request.query.page as string) || 1;
+    const limit = parseInt(request.query.limit as string) || 10;
     const recipeName = request.query.title as string;
-  
+
     const [likedRecipes, savedRecipes, visitedRecipes] = await Promise.all([
       recipeService.getAllLikedRecipesByUserId(user.id),
       recipeService.getAllSavedRecipesByUserId(user.id),
       recipeService.getAllVisitedRecipesByUserId(user.id)
     ]);
-  
+
     const recipes = [...likedRecipes, ...savedRecipes, ...visitedRecipes];
-  
+
     let keywords: string[] = [];
-  
+
     let recipeWords = recipes.map(recipe => {
       return new Set(recipe.title.split(' '));
     });
-  
+
     if (recipeWords.length > 0) {
       keywords = [...recipeWords[0]].filter(word => {
         return recipeWords.every(set => set.has(word));
       });
     }
-  
+
     function containsWord(recipe: Recipe) {
       for (let word of keywords) {
         if (recipe.title.includes(word)) {
@@ -159,9 +177,9 @@ class RecipeController {
       };
       return false;
     };
-  
+
     const recipesIds = recipes.map(recipe => recipe.id);
-  
+
     const [likes, views, comments, findedRecipes] = await Promise.all([
       likeService.getLikesByRecipeIds(recipesIds),
       recipeService.getRecipeVisitsByRecipeIds(recipesIds),
@@ -170,38 +188,38 @@ class RecipeController {
     ]);
 
     findedRecipes.filter(recipe => recipe.ownerId !== user.id);
-  
+
     function calculateScore(recipeId: string) {
-      const likesCount    = likes.filter(like => like.recipeId === recipeId).length;
-      const viewsCount    = views.filter(view => view.recipeId === recipeId).length;
+      const likesCount = likes.filter(like => like.recipeId === recipeId).length;
+      const viewsCount = views.filter(view => view.recipeId === recipeId).length;
       const commentsCount = comments.filter(comment => comment.recipeId === recipeId).length;
-    
+
       return 1 * commentsCount + 2 * viewsCount + 3 * likesCount;
     };
-  
+
     findedRecipes.sort((a, b) => {
       const aScore = calculateScore(a.id);
       const bScore = calculateScore(b.id);
       const aContainsWord = containsWord(a);
       const bContainsWord = containsWord(b);
-  
+
       if (aContainsWord && !bContainsWord) return -1;
       if (!aContainsWord && bContainsWord) return 1;
       return bScore - aScore;
     });
-  
+
     const startIndex = (page - 1) * limit;
-    const endIndex   = startIndex + limit;
-  
+    const endIndex = startIndex + limit;
+
     const paginatedRecipes = findedRecipes.slice(startIndex, endIndex);
-  
+
     response.send(paginatedRecipes.map(recipe => new RecipePreviewDTO(recipe)));
   };
 
   public async getVisitedRecipes(request: Request, response: Response) {
-    const user           = request.user as User;
-    const page           = parseInt(request.query.page as string) || 1;
-    const limit          = parseInt(request.query.limit as string ) || 10;
+    const user = request.user as User;
+    const page = parseInt(request.query.page as string) || 1;
+    const limit = parseInt(request.query.limit as string) || 10;
 
     const visitedRecipes = await recipeService.getVisitedRecipesByUserId(user.id, page, limit);
 
@@ -209,16 +227,16 @@ class RecipeController {
   };
 
   public async createSteps(request: Request, response: Response) {
-    const data     = request.body as Omit<ICreateStepSchema, "recipeId">[];
+    const data = request.body as Omit<ICreateStepSchema, "recipeId">[];
     const recipeId = request.params.recipeId as string;
-    const steps    = await stepService.createStep(data.map(step => ({ ...step, recipeId })));
+    const steps = await stepService.createStep(data.map(step => ({ ...step, recipeId })));
 
     response.send(steps);
   };
 
   public async getSteps(request: Request, response: Response) {
     const recipeId = request.params.recipeId as string;
-    const steps    = await stepService.getStepsByRecipeId(recipeId);
+    const steps = await stepService.getStepsByRecipeId(recipeId);
 
     response.send(steps);
   };
@@ -227,22 +245,22 @@ class RecipeController {
     const user = request.user as User;
     const data = request.body as IUpdateStepSchema[];
 
-    const stepIds        = data.map(step => step.stepId);
-    const step           = await stepService.getStepsByIds(stepIds);
+    const stepIds = data.map(step => step.stepId);
+    const step = await stepService.getStepsByIds(stepIds);
     const stepRecipesIds = step.map(step => step.recipeId);
 
     const recipes = await recipeService.getRecipesByIds(stepRecipesIds);
 
     if (step === null) {
       return response.status(404).send({
-        code   : "step-not-found",
+        code: "step-not-found",
         message: "Step not found!",
       });
     }
 
     if (!recipes.map(recipe => recipe.ownerId.toString()).includes(user.id.toString())) {
       return response.status(403).send({
-        code   : "have-no-access",
+        code: "have-no-access",
         message: "You have no access to change it!",
       });
     }
@@ -253,14 +271,14 @@ class RecipeController {
   };
 
   public async deleteStep(request: Request, response: Response) {
-    const user   = request.user as User;
+    const user = request.user as User;
     const stepId = request.params.stepId as string;
 
     const step = await stepService.getStepById(stepId);
 
     if (step === null) {
       return response.status(404).send({
-        code   : "step-not-found",
+        code: "step-not-found",
         message: "Step not found!"
       });
     }
@@ -269,14 +287,14 @@ class RecipeController {
 
     if (recipe === null) {
       return response.status(404).send({
-        code   : "recipe-not-found",
+        code: "recipe-not-found",
         message: "Recipe not found!"
       })
     }
 
     if (user.id.toString() !== recipe.ownerId.toString()) {
       return response.status(403).send({
-        code   : "have-no-access",
+        code: "have-no-access",
         message: "You have no access to change it!",
       });
     }
@@ -287,9 +305,9 @@ class RecipeController {
   };
 
   public async getSavedRecipes(request: Request, response: Response) {
-    const user  = request.user as User;
-    const page  = parseInt(request.query.page as string) || 1;
-    const limit = parseInt(request.query.limit as string ) || 10;
+    const user = request.user as User;
+    const page = parseInt(request.query.page as string) || 1;
+    const limit = parseInt(request.query.limit as string) || 10;
 
     const savedRecipes = await recipeService.getSavedRecipesByUserId(user.id, page, limit);
 
@@ -297,8 +315,8 @@ class RecipeController {
   };
 
   public async getPopularRecipes(request: Request, response: Response): Promise<void> {
-    const page       = parseInt(request.query.page as string) || 1;
-    const limit      = parseInt(request.query.limit as string ) || 10;
+    const page = parseInt(request.query.page as string) || 1;
+    const limit = parseInt(request.query.limit as string) || 10;
     const recipeName = request.query.title as string;
 
     const recipes = await recipeService.getAllRecipes(recipeName !== 'undefined' ? recipeName : undefined);
@@ -312,17 +330,17 @@ class RecipeController {
     ]);
 
     function calculateScore(recipeId: string): number {
-      const likesCount    = likes.filter(like => like.recipeId === recipeId).length;
-      const viewsCount    = views.filter(view => view.recipeId === recipeId).length;
+      const likesCount = likes.filter(like => like.recipeId === recipeId).length;
+      const viewsCount = views.filter(view => view.recipeId === recipeId).length;
       const commentsCount = comments.filter(comment => comment.recipeId === recipeId).length;
-    
+
       return 1 * commentsCount + 2 * viewsCount + 3 * likesCount;
     };
 
     recipes.sort((a: Recipe, b: Recipe) => calculateScore(b.id) - calculateScore(a.id));
 
     const startIndex = (page - 1) * limit;
-    const endIndex   = startIndex + limit;
+    const endIndex = startIndex + limit;
 
     const paginatedRecipes = recipes.slice(startIndex, endIndex);
 
@@ -330,8 +348,8 @@ class RecipeController {
   };
 
   public async getUserRecipes(request: Request, response: Response) {
-    const userId  = request.params.userId;
-    const sortBy  = request.query.sortBy as 'asc' | 'desc';
+    const userId = request.params.userId;
+    const sortBy = request.query.sortBy as 'asc' | 'desc';
     const recipes = await recipeService.getRecipesByUserIdWithSort(userId, sortBy);
 
     const recipesDTO = recipes.map(recipe => new RecipePreviewDTO(recipe));
